% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry_calc.R
\name{BasalArea}
\alias{BasalArea}
\title{BasalArea}
\usage{
BasalArea(DBH)
}
\arguments{
\item{DBH}{The width of the tree trunk at 1.3 m height measured in cm}

\item{distx}{The distance between the camera and the tree trunk measured in m}
}
\value{
The cross-sectional area of the tree trunk in cm^2
}
\description{
Calculates the Basal Area of each tree in cm^2
}
\examples{
BA1 <- BasalArea(1.45);
}
