% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry_calc.R
\name{PhotoDBH}
\alias{PhotoDBH}
\title{PhotoDBH}
\usage{
PhotoDBH(pixels, distx)
}
\arguments{
\item{pixels}{The number of pixels occupied by the image of a tree trunk at 1.3 m height}

\item{distx}{The distance between the camera and the tree trunk measured in m}
}
\value{
The diameter of the tree in cm
}
\description{
Convert photo pixel counts to Diameter at Breast Height (DBH)
}
\examples{
diam1 <- PhotoDBH(50, 1.2);
}
