% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry_calc.R
\name{TreeHeight}
\alias{TreeHeight}
\title{TreeHeight}
\usage{
TreeHeight(distz, angle)
}
\arguments{
\item{distz}{The distance between the observers eye and the top of the tree in m}

\item{angle}{The angle between the horizontal and the top of the tree from the observer in degrees}
}
\value{
The height of the tree in m
}
\description{
Calculates the Height of each tree in m
}
\examples{
Height1 <- TreeHeight(4.45, 47);
}
